+++ tests/core/test_monolayer.py
@@ -1,4 +1,18 @@
-from tyssue.generation import hexa_grid2d, from_2d_voronoi
-from scipy.spatial import Voronoi
+from tyssue.core.generation import extrude, three_faces_sheet
+from tyssue import Monolayer, config
 
-import tyssue
+
+def test_copy():
+
+    datasets, specs = three_faces_sheet()
+    extruded = extrude(datasets, method='translation')
+    mono = Monolayer('test', extruded,
+                     config.geometry.bulk_spec())
+    assert mono.Nc == 3
+    assert mono.Nf == 24
+    assert mono.Ne == 108
+    assert mono.Nv == 26
+
+    mono2 = mono.copy()
+
+    assert mono2.Nc == 3
